sync-github-releases: extract the per-package sync into its own module
sync.ts mixed two altitudes: the run driver (parse CLI, pick packages,
build the shared context, loop) and the per-package algorithm (read the
package, list its releases, plan, apply).

Move the algorithm — syncPackage, readPackageReleaseNotes and the
SyncContext it consumes — into sync/sync-package.ts, alongside the plan/
apply/tag-scheme steps it orchestrates. sync.ts is now just the driver.
