+++ .github/workflows/sync-github-releases/sync.ts
@@ -3,14 +3,9 @@
 
 main()
 
-import { readFile } from 'node:fs/promises'
-import path from 'node:path'
-import { applyReleasePlan } from './sync/apply-release-plan.ts'
-import { getReleasePlan } from './sync/release-plan.ts'
-import { getTagScheme } from './sync/tag-scheme.ts'
-import { getReleaseNotesByVersion, type ReleaseNotesByVersion } from './utils/changelog.ts'
+import { syncPackage, type SyncContext } from './sync/sync-package.ts'
 import { getPushedFiles, getRepoRoot, getTrackedChangelogFiles, toPackageDirs } from './utils/git.ts'
-import { createReleasesClient, type ReleasesClient } from './utils/github.ts'
+import { createReleasesClient } from './utils/github.ts'
 import { getDefaultBranch, getGithubToken, getRepository } from './utils/github-env.ts'
 
 async function main(): Promise<void> {
@@ -32,14 +27,13 @@ async function main(): Promise<void> {
     console.log('No CHANGELOG.md changes detected — nothing to sync.')
     return
   }
-  const hasMultiplePackages = allPackageDirs.length > 1
 
   const { owner, repo } = getRepository()
   const defaultBranch = getDefaultBranch()
   const context: SyncContext = {
     client: createReleasesClient({ owner, repo, token: getGithubToken() }),
     defaultBranch,
-    hasMultiplePackages,
+    hasMultiplePackages: allPackageDirs.length > 1,
     dryRun,
     // The base of the GitHub web (blob) URL each release's CHANGELOG.md footer links back to —
     // github.com is the web host, distinct from the API the client talks to.
@@ -50,44 +44,6 @@ async function main(): Promise<void> {
     console.log(`Syncing GitHub releases for package directory: ${packageDir}`)
     await syncPackage(packageDir, context)
   }
-}
-
-// The loop-invariant inputs of a sync run, built once and shared across packages. owner/repo aren't
-// here: the client already binds them, and the changelog URL base bakes them in.
-type SyncContext = {
-  client: ReleasesClient
-  defaultBranch: string
-  hasMultiplePackages: boolean
-  dryRun: boolean
-  changelogUrlBase: string
-}
-
-async function syncPackage(packageDir: string, ctx: SyncContext): Promise<void> {
-  // The releases the package's CHANGELOG.md (the source of truth) says should exist …
-  const { packageName, releaseNotesByVersion } = await readPackageReleaseNotes(packageDir, ctx.changelogUrlBase)
-  // … reconciled against the releases currently on GitHub, using this package's tag scheme to match
-  // and name release tags.
-  const githubReleases = await ctx.client.list()
-  const tagScheme = getTagScheme(packageName, ctx.hasMultiplePackages)
-  const plan = getReleasePlan({ githubReleases, releaseNotesByVersion, tagScheme })
-  await applyReleasePlan({ plan, client: ctx.client, defaultBranch: ctx.defaultBranch, dryRun: ctx.dryRun })
-}
-
-// Read a package's identity (its package.json `name`) and the release notes derived from its
-// CHANGELOG.md — keyed by version, each carrying a footer that links back to the changelog on GitHub.
-async function readPackageReleaseNotes(
-  packageDir: string,
-  changelogUrlBase: string,
-): Promise<{ packageName: string; releaseNotesByVersion: ReleaseNotesByVersion }> {
-  const packageDirPath = path.join(process.cwd(), packageDir)
-  const { name: packageName } = JSON.parse(await readFile(path.join(packageDirPath, 'package.json'), 'utf8')) as {
-    name: string
-  }
-  const changelog = await readFile(path.join(packageDirPath, 'CHANGELOG.md'), 'utf8')
-
-  // path.posix.join keeps the URL clean when packageDir is '.' (a repo-root CHANGELOG.md): no `/./`.
-  const changelogUrl = `${changelogUrlBase}/${path.posix.join(packageDir, 'CHANGELOG.md')}`
-  return { packageName, releaseNotesByVersion: getReleaseNotesByVersion(changelog, changelogUrl) }
 }
 
 // Which package directories to sync, in priority order:
